+++ app/Views/page/petaCekKesesuaian.php
@@ -1390,7 +1390,6 @@
                     coordinates.push([longitudeInput + ' ' + latitudeInput]);
                     jsonCoordinates.push([longitudeInput, latitudeInput]);
                 }
-
             });
             console.log(jsonCoordinates);
             // console.log('Nilai Koordinat:', coordinates);
@@ -1435,7 +1434,7 @@
                 prosesDetectInput(jsonCoordinates, "line", geojsonFeature);
             }
         });
-
+        let selectedCounter;
         $("#next_step_byFile").click(function(e) {
             console.log("KLIK");
             if (selectedCounter < 2) {
@@ -1550,240 +1549,7 @@
             undefinedHTML: '[Posisi Koordinat X,Y]'
         });
 
-        // Meass tool
-        map.addControl(mousePositionControl);
-        const measureType = 'Polygon';
-        const style = new ol.style.Style({
-            fill: new ol.style.Fill({
-                color: 'rgba(255, 255, 255, 0.2)',
-            }),
-            stroke: new ol.style.Stroke({
-                color: 'rgba(0, 0, 0, 0.5)',
-                lineDash: [10, 10],
-                width: 2,
-            }),
-            image: new ol.style.Circle({
-                radius: 5,
-                stroke: new ol.style.Stroke({
-                    color: 'rgba(0, 0, 0, 0.7)',
-                }),
-                fill: new ol.style.Fill({
-                    color: 'rgba(255, 255, 255, 0.2)',
-                }),
-            }),
-        });
-        const labelStyle = new ol.style.Style({
-            text: new ol.style.Text({
-                font: '14px Calibri,sans-serif',
-                fill: new ol.style.Fill({
-                    color: 'rgba(255, 255, 255, 1)',
-                }),
-                backgroundFill: new ol.style.Fill({
-                    color: 'rgba(0, 0, 0, 0.7)',
-                }),
-                padding: [3, 3, 3, 3],
-                textBaseline: 'bottom',
-                offsetY: -15,
-            }),
-            image: new ol.style.RegularShape({
-                radius: 8,
-                points: 3,
-                angle: Math.PI,
-                displacement: [0, 10],
-                fill: new ol.style.Fill({
-                    color: 'rgba(0, 0, 0, 0.7)',
-                }),
-            }),
-        });
-        const tipStyle = new ol.style.Style({
-            text: new ol.style.Text({
-                font: '12px Calibri,sans-serif',
-                fill: new ol.style.Fill({
-                    color: 'rgba(255, 255, 255, 1)',
-                }),
-                backgroundFill: new ol.style.Fill({
-                    color: 'rgba(0, 0, 0, 0.4)',
-                }),
-                padding: [2, 2, 2, 2],
-                textAlign: 'left',
-                offsetX: 15,
-            }),
-        });
-        const modifyStyle = new ol.style.Style({
-            image: new ol.style.Circle({
-                radius: 5,
-                stroke: new ol.style.Stroke({
-                    color: 'rgba(0, 0, 0, 0.7)',
-                }),
-                fill: new ol.style.Fill({
-                    color: 'rgba(0, 0, 0, 0.4)',
-                }),
-            }),
-            text: new ol.style.Text({
-                text: 'Drag to modify',
-                font: '12px Calibri,sans-serif',
-                fill: new ol.style.Fill({
-                    color: 'rgba(255, 255, 255, 1)',
-                }),
-                backgroundFill: new ol.style.Fill({
-                    color: 'rgba(0, 0, 0, 0.7)',
-                }),
-                padding: [2, 2, 2, 2],
-                textAlign: 'left',
-                offsetX: 15,
-            }),
-        });
-        const segmentStyle = new ol.style.Style({
-            text: new ol.style.Text({
-                font: '12px Calibri,sans-serif',
-                fill: new ol.style.Fill({
-                    color: 'rgba(255, 255, 255, 1)',
-                }),
-                backgroundFill: new ol.style.Fill({
-                    color: 'rgba(0, 0, 0, 0.4)',
-                }),
-                padding: [2, 2, 2, 2],
-                textBaseline: 'bottom',
-                offsetY: -12,
-            }),
-            image: new ol.style.RegularShape({
-                radius: 6,
-                points: 3,
-                angle: Math.PI,
-                displacement: [0, 8],
-                fill: new ol.style.Fill({
-                    color: 'rgba(0, 0, 0, 0.4)',
-                }),
-            }),
-        });
-        const segmentStyles = [segmentStyle];
-        // Sumber data untuk menyimpan fitur pengukuran
-        const measureSource = new ol.source.Vector();
-        const modify = new ol.interaction.Modify({
-            source: measureSource,
-            style: modifyStyle
-        });
-        let tipPoint;
-
-        function styleFunction(feature, segments, drawType, tip) {
-            const styles = [];
-            const geometry = feature.getGeometry();
-            const type = geometry.getType();
-            let point, label, line;
-            if (!drawType || drawType === type || type === 'Point') {
-                styles.push(style);
-                if (type === 'Polygon') {
-                    point = geometry.getInteriorPoint();
-                    label = formatArea(geometry);
-                    line = new ol.geom.LineString(geometry.getCoordinates()[0]);
-                } else if (type === 'LineString') {
-                    point = new ol.geom.Point(geometry.getLastCoordinate());
-                    label = formatLength(geometry);
-                    line = geometry;
-                }
-            }
-            if (segments && line) {
-                let count = 0;
-                line.forEachSegment(function(a, b) {
-                    const segment = new ol.geom.LineString([a, b]);
-                    const label = formatLength(segment);
-                    if (segmentStyles.length - 1 < count) {
-                        segmentStyles.push(segmentStyle.clone());
-                    }
-                    const segmentPoint = new ol.geom.Point(segment.getCoordinateAt(0.5));
-                    segmentStyles[count].setGeometry(segmentPoint);
-                    segmentStyles[count].getText().setText(label);
-                    styles.push(segmentStyles[count]);
-                    count++;
-                });
-            }
-            if (label) {
-                labelStyle.setGeometry(point);
-                labelStyle.getText().setText(label);
-                styles.push(labelStyle);
-            }
-            if (
-                tip &&
-                type === 'Point' &&
-                !modify.getOverlay().getSource().getFeatures().length
-            ) {
-                tipPoint = geometry;
-                tipStyle.getText().setText(tip);
-                styles.push(tipStyle);
-            }
-            return styles;
-        }
-        const formatLength = function(line) {
-            const length = ol.sphere.getLength(line);
-            let output;
-            if (length > 100) {
-                output = Math.round((length / 1000) * 100) / 100 + ' km';
-            } else {
-                output = Math.round(length * 100) / 100 + ' m';
-            }
-            return output;
-        };
-        const formatArea = function(polygon) {
-            const area = ol.sphere.getArea(polygon);
-            let output;
-            if (area > 10000) {
-                output = Math.round((area / 1000000) * 100) / 100 + ' km\xB2';
-            } else {
-                output = Math.round(area * 100) / 100 + ' m\xB2';
-            }
-            return output;
-        };
-        const measure = new ol.layer.Vector({
-            source: measureSource,
-            style: function(feature) {
-                return styleFunction(feature, "checked");
-            },
-        });
-        map.addLayer(measure);
-        let draw;
-        // Fungsi untuk menambahkan interaksi pengukuran polyline
-        function addMeasurement() {
-            map.getViewport().style.cursor = "default";
-            const drawType = measureType; // Tipe pengukuran diubah menjadi LineString
-            const activeTip =
-                'Click to continue drawing the ' +
-                (drawType === 'Polygon' ? 'polygon' : 'line');
-            const idleTip = 'Click to start measuring';
-            let tip = idleTip; // Tip awal saat pengukuran dimulai
-            draw = new ol.interaction.Draw({
-                source: measureSource,
-                type: drawType,
-                style: function(feature) {
-                    return styleFunction(feature, "checked", drawType, tip);
-                },
-            });
-            draw.on('drawstart', function() {
-                tip = activeTip;
-            });
-            draw.on('drawend', function() {
-                map.getViewport().style.cursor = "grab";
-                modifyStyle.setGeometry(tipPoint);
-                map.removeInteraction(draw);
-                tip = idleTip;
-            });
-            map.addInteraction(draw);
-        }
-
-        // Buat tombol kontrol Meass
-        var rulerControl = new ol.control.Control({
-            element: document.getElementById('ruler-button'),
-        });
-        map.addControl(rulerControl);
-        $(rulerControl.element).click(function(e) {
-            map.removeInteraction(draw);
-            measureSource.clear();
-            addMeasurement();
-            tip = 'Click to start measuring';
-        });
-
-
         $('#isiByFile').change(function(e) {
-            let selectedCounter;
             const file = e.target.files[0];
             console.log(file);
             const reader = new FileReader();
@@ -1905,6 +1671,7 @@
         });
 
         function geojsonFromFile(geojson) {
+            geojsonFeature = [];
             const type = geojson.features[0].geometry.type;
             jsonCoordinates = geojson.features[0].geometry.coordinates[0];
             // console.log(jsonCoordinates);
@@ -1914,12 +1681,14 @@
                 selectedCounter = 3;
             }
             if (selectedCounter < 2) {
+                geojsonFeature = turf.point([jsonCoordinates[0][0], jsonCoordinates[0][1]]);
                 var pointFeature = new ol.Feature({
                     geometry: new ol.geom.Point(ol.proj.fromLonLat(jsonCoordinates[0]))
                 });
                 vectorSource.addFeature(pointFeature);
                 styleDraw = markerStyle;
             } else {
+                geojsonFeature = turf.polygon([jsonCoordinates]);
                 var polygonFeature = new ol.Feature({
                     geometry: new ol.geom.Polygon([jsonCoordinates.map(coordinate => ol.proj.transform(coordinate, 'EPSG:4326', 'EPSG:3857'))])
                 });
@@ -1933,16 +1702,7 @@
             }, '<?= base_url('/data/petaPreview'); ?>');
         }
     </script>
-
-    <!-- <script>
-        var a = "Pipa Minyak dan Gas";
-        var b = "Kabel Telekomunikasi";
-        var c = "Pipa Minyak dan Gas";
-        var x = a === b;
-        var z = a === c;
-        console.log(x);
-        console.log(z);
-    </script> -->
+    <script src="/js/meass-tool.js"></script>
 
 </body>
