+++ public/dokumen/upload-dokumen/map_651946471465f.json
@@ -1,0 +1,24 @@
+{
+  "type": "FeatureCollection",
+  "features": [
+    {
+      "type": "Feature",
+      "properties": {},
+      "geometry": {
+        "coordinates": [
+          [
+            [116.9487864213599, -1.3703618016229626],
+            [116.74030204596897, -1.5301485708451281],
+            [116.89319058792353, -1.6412976384698652],
+            [117.1989676718294, -1.4606774520233614],
+            [117.28931090116691, -1.148032310289679],
+            [117.23371506772895, -0.9743251949971352],
+            [117.15032131757198, -1.113291680401602],
+            [116.9487864213599, -1.3703618016229626]
+          ]
+        ],
+        "type": "Polygon"
+      }
+    }
+  ]
+}
